Refactor masonry grid
